Refactor: 좋아요 컴포넌트 분리
