+++ src/components/DetailSectionTop/DetailToursInfo.tsx
@@ -1,7 +1,4 @@
-import { HeartIcon } from '@components/common/icons/Icons';
-import { postLikedTours, deleteLikedTours } from '@api/tours';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
+import Like from '@components/common/like/Like';
 
 interface DetailToursInfoProps {
   infoData: tourDetail;
@@ -9,39 +6,6 @@ interface DetailToursInfoProps {
 
 export default function DetailToursInfo({ infoData }: DetailToursInfoProps) {
   const { title, liked, originalThumbnailUrl, id } = infoData;
-
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-
-  const { mutate: likeMutate } = useMutation({
-    mutationFn: (id: number) => postLikedTours({ id }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['details'] });
-    },
-    onError: () => console.log('error'),
-  });
-
-  const { mutate: unlikeMutate } = useMutation({
-    mutationFn: (id: number) => deleteLikedTours({ id }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['details'] });
-    },
-    onError: () => console.log('error'),
-  });
-
-  const onClickLikeButton = () => {
-    const token = localStorage.getItem('accessToken');
-    if (!token) {
-      // 비로그인 알람창 처리 필요
-      navigate('/signin');
-    } else {
-      if (liked === false) {
-        likeMutate(id);
-      } else {
-        unlikeMutate(id);
-      }
-    }
-  };
 
   return (
     <>
@@ -56,13 +20,7 @@ export default function DetailToursInfo({ infoData }: DetailToursInfoProps) {
         <h1 className="font-['Pretendard'] text-2xl font-bold text-black ">
           {title}
         </h1>
-        <div className="top-75 h-[24px] w-[24px] cursor-pointer">
-          <HeartIcon
-            fill={liked ? '#FF2167' : '#D7D7D7'}
-            color="none"
-            onClick={onClickLikeButton}
-          />
-        </div>
+        <Like liked={liked} id={id} />
       </div>
     </>
   );
